+++ src/test/scala/chisel3/tests/AsyncResetUsingBlackBoxTest.scala
@@ -13,7 +13,7 @@ import treadle.BlackBoxFactoriesAnnotation
   * circuit that illustrates usage of async register
   * @param resetValue value on reset
   */
-class AsyncResetRegModule(resetValue: Int) extends Module {
+class AsyncResetUsingBlackBoxModule(resetValue: Int) extends Module {
   //noinspection TypeAnnotation
   val io = IO(new Bundle {
     val in = Input(UInt(1.W))
@@ -30,7 +30,7 @@ class AsyncResetRegModule(resetValue: Int) extends Module {
   io.out := reg.q
 }
 
-class AsyncResetFeedbackModule() extends Module {
+class AsyncResetUsingBlackBoxFeedbackModule() extends Module {
   //noinspection TypeAnnotation
   val io = IO(new Bundle {
     val out = Output(Vec(2, UInt(1.W)))
@@ -53,21 +53,21 @@ class AsyncResetFeedbackModule() extends Module {
   reg1.en := 1.U
 }
 
-class AsyncResetRegTest extends FreeSpec with ChiselScalatestTester {
+class AsyncResetUsingBlackBoxTest extends FreeSpec with ChiselScalatestTester {
   private val annotations = Seq(BlackBoxFactoriesAnnotation(Seq(new AsyncResetBlackBoxFactory)))
 
   "register is zero, after tester startup's default reset" in {
-    test(new AsyncResetRegModule(0)).withAnnotations(annotations) { dut =>
+    test(new AsyncResetUsingBlackBoxModule(0)).withAnnotations(annotations) { dut =>
       dut.io.out.expect(0.U)
     }
   }
   "register is one, after tester startup's default reset" in {
-    test(new AsyncResetRegModule(1)).withAnnotations(annotations) { dut =>
+    test(new AsyncResetUsingBlackBoxModule(1)).withAnnotations(annotations) { dut =>
       dut.io.out.expect(1.U)
     }
   }
   "reset a register works after register has been altered" in {
-    test(new AsyncResetRegModule(1)).withAnnotations(annotations) { dut =>
+    test(new AsyncResetUsingBlackBoxModule(1)).withAnnotations(annotations) { dut =>
       // The register starts at 1 after default reset in tester startup
       dut.io.out.expect(1.U)
 
@@ -99,7 +99,7 @@ class AsyncResetRegTest extends FreeSpec with ChiselScalatestTester {
     }
   }
   "de-assert reset behaviour" in {
-    test(new AsyncResetRegModule(1)).withAnnotations(annotations) { dut =>
+    test(new AsyncResetUsingBlackBoxModule(1)).withAnnotations(annotations) { dut =>
       // register is reset at startup, and set to zero by poking
       dut.clock.step()
       dut.io.out.expect(0.U)
@@ -121,7 +121,7 @@ class AsyncResetRegTest extends FreeSpec with ChiselScalatestTester {
     }
   }
   "feedback into itself" in {
-    test(new AsyncResetFeedbackModule).withAnnotations(annotations) { dut =>
+    test(new AsyncResetUsingBlackBoxFeedbackModule).withAnnotations(annotations) { dut =>
       dut.io.out(0).expect(0.U)
       dut.io.out(1).expect(1.U)
